feat: better divider types
