+++ packages/divider/index.ts
@@ -1,2 +1,2 @@
 export { default as Divider } from './Divider'
-export * from './types'
+export * from './Divider'
